backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sentence-transformers==3.0.1
numpy==2.0.1
pillow==10.4.0
